refactor(home): split home_page.dart god file into part extensions (stage 2, slice 6)
Completes the Stage-2 decomposition. _HomePageState was ~2,240 lines in
one file; its methods are now split across five part files, each a
private extension on _HomePageState (part of the same library, so they
keep full access to page state — verbatim moves, zero param threading):

- home/home_page_chrome.dart   — mode toggle + status bar build helpers
- home/home_page_handlers.dart — tap/context-menu/folder/toolbar callbacks
- home/home_page_dialogs.dart  — folder + delete dialogs, move-to-folder
- home/home_page_char_ops.dart — character/BYAF import + bulk progress
- home/home_page_transfer.dart — character/group export + group import

home_page.dart keeps the State fields, lifecycle (initState/dispose/
didChangeDependencies), build(), the small toggles/cache helpers, and a
one-line applyState() wrapper — setState is @protected and the analyzer
doesn't treat extension methods as instance members of the State
subclass, so the parts rebuild through applyState(). Extensions are named
privately to keep the private State type out of any public API.

home_page.dart: 2494 → 423. Every home file is now under the 500-line cap
except character_grid_card.dart (599, one cohesive card — approved).
Full analyze clean (only pre-existing deprecation infos in untouched
files); UI suite green; the sole golden diff (grid_cards, 27px/0.00%) is
the pre-existing environment-AA noise, identical to baseline.
