+++ lib/ui/pages/home/home_page_handlers.dart
@@ -1,0 +1,443 @@
+// Copyright (C) 2026 Front Porch AI
+// SPDX-License-Identifier: AGPL-3.0-or-later
+//
+// This file is part of Front Porch AI.
+//
+// Front Porch AI is free software: you can redistribute it and/or modify
+// it under the terms of the GNU Affero General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// Front Porch AI is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+// GNU Affero General Public License for more details.
+//
+// You should have received a copy of the GNU Affero General Public License
+// along with Front Porch AI. If not, see <https://www.gnu.org/licenses/>.
+
+part of '../home_page.dart';
+
+/// Tap, context-menu, folder-navigation and toolbar-callback handlers.
+///
+/// Split out of the _HomePageState god file as a private extension
+/// (part of the same library, so it keeps full access to page state).
+extension _HomePageHandlers on _HomePageState {
+
+  void _confirmDeleteGroup(BuildContext context, GroupChat group) {
+    showDialog(
+      context: context,
+      builder: (ctx) => AlertDialog(
+        backgroundColor: const Color(0xFF2D1111),
+        shape: RoundedRectangleBorder(
+          borderRadius: BorderRadius.circular(16),
+          side: const BorderSide(color: Colors.redAccent, width: 2),
+        ),
+        title: const Text(
+          'Delete Group',
+          style: TextStyle(
+            color: Colors.redAccent,
+            fontWeight: FontWeight.bold,
+          ),
+        ),
+        content: Text(
+          'Delete group "${group.name}"?\n\nThe characters themselves will NOT be deleted.',
+          style: const TextStyle(color: Colors.white70, height: 1.5),
+        ),
+        actions: [
+          TextButton(
+            onPressed: () => Navigator.pop(ctx),
+            child: const Text(
+              'Cancel',
+              style: TextStyle(color: Colors.white54),
+            ),
+          ),
+          ElevatedButton(
+            style: ElevatedButton.styleFrom(backgroundColor: Colors.redAccent),
+            onPressed: () async {
+              Navigator.of(ctx).pop();
+              final groupRepo = Provider.of<GroupChatRepository>(
+                context,
+                listen: false,
+              );
+              await groupRepo.delete(group.id);
+              // No post-delete snackbar for groups (character delete shows one via the outer context)
+            },
+            child: const Text('Delete'),
+          ),
+        ],
+      ),
+    );
+  }
+
+  // ─── Group Creation Dialog ──────────────────────────────────────
+
+  void _showMoveToFolderDialog(
+    BuildContext context,
+    CharacterRepository repo,
+    FolderService folderService,
+  ) {
+    final folders = folderService.folders.toList()
+      ..sort((a, b) {
+        final pathA = folderService.getFolderPath(a.id).toLowerCase();
+        final pathB = folderService.getFolderPath(b.id).toLowerCase();
+        return pathA.compareTo(pathB);
+      });
+
+    showDialog(
+      context: context,
+      builder: (ctx) => AlertDialog(
+        backgroundColor: const Color(0xFF1E293B),
+        shape: RoundedRectangleBorder(
+          borderRadius: BorderRadius.circular(16),
+          side: const BorderSide(color: Colors.blueAccent, width: 0.5),
+        ),
+        title: Row(
+          children: [
+            const Icon(Icons.drive_file_move, color: Colors.blueAccent),
+            const SizedBox(width: 12),
+            Text(
+              'Move ${_selectedCharacterIds.length} character${_selectedCharacterIds.length == 1 ? '' : 's'} to folder',
+              style: const TextStyle(fontSize: 16, fontWeight: FontWeight.bold),
+            ),
+          ],
+        ),
+        content: SizedBox(
+          width: 320,
+          child: SingleChildScrollView(
+            child: Column(
+              mainAxisSize: MainAxisSize.min,
+              children: [
+                if (folders.isEmpty)
+                  const Padding(
+                    padding: EdgeInsets.symmetric(vertical: 12),
+                    child: Text(
+                      'No folders yet. Create one below.',
+                      style: TextStyle(color: Colors.white54),
+                    ),
+                  ),
+                ...folders.map((folder) {
+                  final folderPath = folderService.getFolderPath(folder.id);
+                  final isSubfolder = folder.parentId != null;
+                  return ListTile(
+                    leading: Icon(
+                      isSubfolder
+                          ? Icons.subdirectory_arrow_right
+                          : Icons.folder,
+                      color: Colors.amberAccent,
+                    ),
+                    title: Text(
+                      folderPath,
+                      style: const TextStyle(color: Colors.white),
+                    ),
+                    subtitle: Text(
+                      '${folder.characterPaths.length} characters',
+                      style: const TextStyle(
+                        color: Colors.white38,
+                        fontSize: 12,
+                      ),
+                    ),
+                    shape: RoundedRectangleBorder(
+                      borderRadius: BorderRadius.circular(8),
+                    ),
+                    hoverColor: Colors.white10,
+                    onTap: () async {
+                      Navigator.pop(ctx);
+                      await _moveSelectedToFolder(
+                        context,
+                        folder.id,
+                        repo,
+                        folderService,
+                      );
+                    },
+                  );
+                }),
+                const Divider(color: Colors.white12),
+                ListTile(
+                  leading: const Icon(
+                    Icons.create_new_folder,
+                    color: Colors.greenAccent,
+                  ),
+                  title: const Text(
+                    'New Folder',
+                    style: TextStyle(color: Colors.greenAccent),
+                  ),
+                  shape: RoundedRectangleBorder(
+                    borderRadius: BorderRadius.circular(8),
+                  ),
+                  hoverColor: Colors.white10,
+                  onTap: () async {
+                    Navigator.pop(ctx);
+                    final name = await _promptFolderName(context);
+                    if (name != null && name.isNotEmpty && context.mounted) {
+                      final folder = await folderService.createFolder(name);
+                      await _moveSelectedToFolder(
+                        context,
+                        folder.id,
+                        repo,
+                        folderService,
+                      );
+                    }
+                  },
+                ),
+              ],
+            ),
+          ),
+        ),
+        actions: [
+          TextButton(
+            onPressed: () => Navigator.pop(ctx),
+            child: const Text(
+              'Cancel',
+              style: TextStyle(color: Colors.white54),
+            ),
+          ),
+        ],
+      ),
+    );
+  }
+
+  Future<String?> _promptFolderName(BuildContext context) async {
+    final controller = TextEditingController();
+    return showDialog<String>(
+      context: context,
+      builder: (ctx) => AlertDialog(
+        backgroundColor: const Color(0xFF1E293B),
+        shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(16)),
+        title: const Text('New Folder'),
+        content: TextField(
+          controller: controller,
+          autofocus: true,
+          decoration: const InputDecoration(hintText: 'Folder name'),
+          onSubmitted: (val) => Navigator.pop(ctx, val),
+        ),
+        actions: [
+          TextButton(
+            onPressed: () => Navigator.pop(ctx),
+            child: const Text(
+              'Cancel',
+              style: TextStyle(color: Colors.white54),
+            ),
+          ),
+          ElevatedButton(
+            onPressed: () => Navigator.pop(ctx, controller.text),
+            style: ElevatedButton.styleFrom(backgroundColor: Colors.blueAccent),
+            child: const Text('Create'),
+          ),
+        ],
+      ),
+    );
+  }
+
+  Future<void> _moveSelectedToFolder(
+    BuildContext context,
+    String folderId,
+    CharacterRepository repo,
+    FolderService folderService,
+  ) async {
+    // Resolve selected IDs back to imagePaths
+    for (final id in _selectedCharacterIds) {
+      final card = repo.characters
+          .where((c) => _getCharacterIdFromCard(c) == id)
+          .firstOrNull;
+      if (card?.imagePath != null) {
+        await folderService.addToFolder(folderId, card!.imagePath!);
+      }
+    }
+    if (context.mounted) {
+      ScaffoldMessenger.of(context).showSnackBar(
+        SnackBar(
+          content: Text(
+            'Moved ${_selectedCharacterIds.length} character${_selectedCharacterIds.length == 1 ? '' : 's'} to folder',
+          ),
+        ),
+      );
+    }
+    _cancelSelection();
+  }
+
+  // (old group creator dialog variables and body fully removed — 2026 overhaul)
+
+  // ─── Folder Actions ─────────────────────────────────────────────
+
+  void _createFolder(
+    BuildContext context,
+    FolderService folderService, {
+    String? parentId,
+  }) {
+    final controller = TextEditingController();
+    showDialog(
+      context: context,
+      builder: (ctx) => AlertDialog(
+        backgroundColor: const Color(0xFF1F2937),
+        title: Text(
+          parentId != null ? 'New Subfolder' : 'New Folder',
+          style: const TextStyle(color: Colors.white),
+        ),
+        content: TextField(
+          controller: controller,
+          autofocus: true,
+          style: const TextStyle(color: Colors.white),
+          decoration: const InputDecoration(
+            hintText: 'Folder name...',
+            hintStyle: TextStyle(color: Colors.white38),
+          ),
+          onSubmitted: (value) {
+            if (value.trim().isNotEmpty) {
+              folderService.createFolder(value.trim(), parentId: parentId);
+              Navigator.pop(ctx);
+            }
+          },
+        ),
+        actions: [
+          TextButton(
+            onPressed: () => Navigator.pop(ctx),
+            child: const Text(
+              'Cancel',
+              style: TextStyle(color: Colors.white54),
+            ),
+          ),
+          ElevatedButton(
+            onPressed: () {
+              if (controller.text.trim().isNotEmpty) {
+                folderService.createFolder(
+                  controller.text.trim(),
+                  parentId: parentId,
+                );
+                Navigator.pop(ctx);
+              }
+            },
+            style: ElevatedButton.styleFrom(
+              backgroundColor: Colors.amber.shade700,
+            ),
+            child: const Text('Create'),
+          ),
+        ],
+      ),
+    );
+  }
+
+  void _renameFolder(
+    BuildContext context,
+    CharacterFolder folder,
+    FolderService folderService,
+  ) {
+    final controller = TextEditingController(text: folder.name);
+    showDialog(
+      context: context,
+      builder: (ctx) => AlertDialog(
+        backgroundColor: const Color(0xFF1F2937),
+        title: const Text(
+          'Rename Folder',
+          style: TextStyle(color: Colors.white),
+        ),
+        content: TextField(
+          controller: controller,
+          autofocus: true,
+          style: const TextStyle(color: Colors.white),
+          onSubmitted: (value) {
+            if (value.trim().isNotEmpty) {
+              folderService.renameFolder(folder.id, value.trim());
+              Navigator.pop(ctx);
+            }
+          },
+        ),
+        actions: [
+          TextButton(
+            onPressed: () => Navigator.pop(ctx),
+            child: const Text(
+              'Cancel',
+              style: TextStyle(color: Colors.white54),
+            ),
+          ),
+          ElevatedButton(
+            onPressed: () {
+              if (controller.text.trim().isNotEmpty) {
+                folderService.renameFolder(folder.id, controller.text.trim());
+                Navigator.pop(ctx);
+              }
+            },
+            style: ElevatedButton.styleFrom(
+              backgroundColor: Colors.amber.shade700,
+            ),
+            child: const Text('Rename'),
+          ),
+        ],
+      ),
+    );
+  }
+
+  void _deleteFolder(
+    BuildContext context,
+    CharacterFolder folder,
+    FolderService folderService,
+  ) {
+    showDialog(
+      context: context,
+      builder: (ctx) => AlertDialog(
+        backgroundColor: const Color(0xFF2D1111),
+        shape: RoundedRectangleBorder(
+          borderRadius: BorderRadius.circular(16),
+          side: const BorderSide(color: Colors.redAccent, width: 2),
+        ),
+        title: const Text(
+          'Delete Folder',
+          style: TextStyle(
+            color: Colors.redAccent,
+            fontWeight: FontWeight.bold,
+          ),
+        ),
+        content: Text(
+          'Delete "${folder.name}"?\n\nCharacters inside will NOT be deleted — they\'ll return to the top level.',
+          style: const TextStyle(color: Colors.white70, height: 1.5),
+        ),
+        actions: [
+          TextButton(
+            onPressed: () => Navigator.pop(ctx),
+            child: const Text(
+              'Cancel',
+              style: TextStyle(color: Colors.white54),
+            ),
+          ),
+          ElevatedButton(
+            style: ElevatedButton.styleFrom(backgroundColor: Colors.redAccent),
+            onPressed: () {
+              folderService.deleteFolder(folder.id);
+              Navigator.pop(ctx);
+              if (_activeFolderId == folder.id) {
+                applyState(() {
+                  if (_folderStack.isNotEmpty) {
+                    _activeFolderId = _folderStack.removeLast();
+                  } else {
+                    _activeFolderId = null;
+                  }
+                });
+              }
+            },
+            child: const Text('Delete'),
+          ),
+        ],
+      ),
+    );
+  }
+
+  // ─── Character Actions ──────────────────────────────────────────
+
+  ButtonStyle _buttonStyle() {
+    return ElevatedButton.styleFrom(
+      backgroundColor: AppColors.resolve(
+        context,
+        Colors.white.withValues(alpha: 0.1),
+        Colors.black.withValues(alpha: 0.05),
+      ),
+      foregroundColor: AppColors.textPrimary(context),
+      padding: const EdgeInsets.symmetric(horizontal: 20, vertical: 16),
+      shape: RoundedRectangleBorder(
+        borderRadius: BorderRadius.circular(12),
+        side: BorderSide(
+          color: AppColors.resolve(context, Colors.white24, Colors.black26),
+        ),
+      ),
+    );
+  }
+}
